health: #184 moved all test files next to source files into root test folder of core lib
