--- packages/core/src/test/entry.ts
@@ -1,2 +1,0 @@
-import "./spec-bundle";
-import "./vendor";
