+++ packages/core/src/test/expansion/intersect-expansion.test.ts
@@ -1,5 +1,5 @@
 import { ExpansionValue } from "@entity-space/common";
-import { Expansion } from "../expansion";
+import { Expansion } from "../../lib/expansion/expansion";
 
 function intersectExpansion(a: ExpansionValue, b: ExpansionValue): boolean | ExpansionValue {
     const intersected = new Expansion(b).intersect(new Expansion(a));
