+++ packages/core/src/test/expansion/reduce-expansion.test.ts
@@ -1,5 +1,5 @@
 import { ExpansionValue } from "@entity-space/common";
-import { Expansion } from "../expansion";
+import { Expansion } from "../../lib/expansion/expansion";
 
 function reduceExpansion(a: ExpansionValue, b: ExpansionValue): boolean | ExpansionValue {
     const reduced = new Expansion(b).reduce(new Expansion(a));
