--- packages/core/src/test/vendor.ts
@@ -1,1 +1,0 @@
-import "core-js";
