+++ packages/core/tsconfig.lib.json
@@ -1,11 +1,10 @@
 {
     "extends": "./tsconfig.json",
     "compilerOptions": {
-        "module": "commonjs",
         "outDir": "../../dist/out-tsc",
         "declaration": true,
-        "types": ["node"]
+        "types": []
     },
-    "exclude": ["**/*.spec.ts", "**/*.test.ts", "**/test/**/*.ts", "jest.config.ts"],
-    "include": ["**/*.ts"]
+    "include": ["**/*.ts"],
+    "exclude": ["**/*.test.ts", "jest.config.ts"]
 }
