+++ packages/criteria/tsconfig.lib.json
@@ -6,5 +6,5 @@
         "types": []
     },
     "include": ["**/*.ts"],
-    "exclude": ["**/*.spec.ts", "jest.config.ts"]
+    "exclude": ["**/*.test.ts", "jest.config.ts"]
 }
